+++ app/src/main/kotlin/org/virtuslab/bazelsteward/app/BazelRulesDependencyKind.kt
@@ -1,11 +1,9 @@
 package org.virtuslab.bazelsteward.app
 
-import org.virtuslab.bazelsteward.core.library.Library
 import org.virtuslab.bazelsteward.core.library.Version
 import org.virtuslab.bazelsteward.core.replacement.BazelRuleHeuristic
 import org.virtuslab.bazelsteward.core.replacement.Heuristic
 import org.virtuslab.bazelsteward.core.rules.RuleLibrary
-import org.virtuslab.bazelsteward.core.rules.RuleLibraryId
 import org.virtuslab.bazelsteward.github.GithubRulesResolver
 import org.virtuslab.bazelsteward.rules.BazelRulesExtractor
 import java.nio.file.Path
@@ -13,14 +11,13 @@ import java.nio.file.Path
 class BazelRulesDependencyKind(
   private val bazelRulesExtractor: BazelRulesExtractor,
   private val githubRulesResolver: GithubRulesResolver
-) : DependencyKind<RuleLibraryId> {
+) : DependencyKind<RuleLibrary> {
   override val name: String = "bazel-rules"
 
-  override suspend fun findAvailableVersions(workspaceRoot: Path): Map<Library, List<Version>> {
+  override suspend fun findAvailableVersions(workspaceRoot: Path): Map<RuleLibrary, List<Version>> {
     val usedBazelRules = bazelRulesExtractor.extractCurrentRules(workspaceRoot)
     return usedBazelRules
       .associateWith { it: RuleLibrary -> githubRulesResolver.resolveRuleVersions(it.id).values.toList() }
-      as Map<Library, List<Version>>
   }
 
   override val defaultSearchPatterns: List<PathPattern> = listOf(
