+++ app/src/main/kotlin/org/virtuslab/bazelsteward/app/MavenDependencyKind.kt
@@ -6,8 +6,8 @@ import org.virtuslab.bazelsteward.core.library.Version
 import org.virtuslab.bazelsteward.core.replacement.Heuristic
 import org.virtuslab.bazelsteward.core.replacement.VersionOnlyHeuristic
 import org.virtuslab.bazelsteward.core.replacement.WholeLibraryHeuristic
+import org.virtuslab.bazelsteward.maven.MavenCoordinates
 import org.virtuslab.bazelsteward.maven.MavenDataExtractor
-import org.virtuslab.bazelsteward.maven.MavenLibraryId
 import org.virtuslab.bazelsteward.maven.MavenRepository
 import java.nio.file.Path
 
@@ -16,15 +16,15 @@ private val logger = KotlinLogging.logger {}
 class MavenDependencyKind(
   private val mavenDataExtractor: MavenDataExtractor,
   private val mavenRepository: MavenRepository
-) : DependencyKind<MavenLibraryId> {
+) : DependencyKind<MavenCoordinates> {
 
   override val name: String = "maven"
 
-  override suspend fun findAvailableVersions(workspaceRoot: Path): Map<Library, List<Version>> {
+  override suspend fun findAvailableVersions(workspaceRoot: Path): Map<MavenCoordinates, List<Version>> {
     val data = mavenDataExtractor.extract()
     logger.debug { "Repositories " + data.repositories.toString() }
     logger.debug { "Dependencies: " + data.dependencies.map { it.id.name + " " + it.version.value }.toString() }
-    return mavenRepository.findVersions(data) as Map<Library, List<Version>>
+    return mavenRepository.findVersions(data)
   }
 
   override val defaultSearchPatterns: List<PathPattern> = listOf(
