+++ setup-bsp.sh
@@ -1,2 +1,2 @@
 #!/usr/bin/env bash
-cs launch -r m2Local org.jetbrains.bsp:bazel-bsp:2.5.0 -M org.jetbrains.bsp.bazel.install.Install
+cs launch -r m2Local org.jetbrains.bsp:bazel-bsp:2.5.1 -M org.jetbrains.bsp.bazel.install.Install -- --bazel-workspace . --directory ../bazel-steward-bsp
